refactor(domain): 불필요한 클래스 삭제
테스트 변경에 따라 불필요해진 인터페이스 LineGenerator 삭제 및 DefaultLineGenerator 클래스명 LineGenerator로 변경
