+++ src/main/java/ladder/domain/LineGenerator.java
@@ -1,5 +1,18 @@
 package ladder.domain;
 
-public interface LineGenerator {
-    Direction generate();
+import static ladder.domain.Direction.NONE;
+import static ladder.domain.Direction.RIGHT;
+
+import java.util.Random;
+
+public class LineGenerator {
+
+    private static final Random random = new Random();
+
+    public Direction generate() {
+        if (random.nextBoolean()) {
+            return RIGHT;
+        }
+        return NONE;
+    }
 }
